Fix ObservableComponent::useEvent ignoring seed
ObservableComponent::useEvent checked whether
more than one argument was supplied (correct), but
used the 3rd argument as a seed value instead.
That argument is always undefined when using
Refract according to the docs.

This bug was introduced with commit d24f0f099b2.

This commit changes the function signature back
to the original one of using 2 args, one being
optional. The check for the actual argument count
is then done using the arguments object.